% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalSize.R
\name{distFS}
\alias{distFS}
\title{disFSfast}
\usage{
distFS(R, s0in, i0in, r0in = NULL)
}
\arguments{
\item{R}{Value of R0}

\item{s0in}{Number of initial susceptibles}

\item{i0in}{Number of initial infectious}

\item{r0in}{Number of initial recovered}
}
\value{
distribution of number of individuals that were infected during the outbreak
}
\description{
this function will first create a generation based table with states and then gives the final size distribution as output
}
