% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.final_size}
\alias{.final_size}
\title{Calculate the final size of an epidemic}
\usage{
.final_size(parameters)
}
\arguments{
\item{parameters}{A named list of parameters for the final size calculation.
See the R function documentation for details and input checking.}
}
\description{
An internal function that interfaces between the R function
\code{final_size()} and functions in the package header.
}
