% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.prepare_data}
\alias{.prepare_data}
\title{Prepare population data for solvers.}
\usage{
.prepare_data(
  contact_matrix,
  demography_vector,
  p_susceptibility,
  susceptibility
)
}
\arguments{
\item{contact_matrix}{Social contact matrix. Entry $mm_{ij}$ gives average
number of contacts in group $i$ reported by participants in group j}

\item{demography_vector}{Demography vector. Entry $pp_{i}$ gives proportion
of total population in group $i$ (model will normalise if needed)}

\item{p_susceptibility}{A matrix giving the probability that an individual
in demography group $i$ is in risk (or susceptibility) group $j$.
Each row represents the overall distribution of individuals in demographic
group $i$ across risk groups, and each row \emph{must sum to 1.0}.}

\item{susceptibility}{A matrix giving the susceptibility of individuals in
demographic group $i$ and risk group $j$.}
}
\value{
A list object with named elements: "contact_matrix",
"demography_vector", "p_susceptibility_", and "susceptibility".
The contact matrix is replicated row and column wise for each risk group
and the demography vector is replicated for each risk group.
}
\description{
An internal function that prepares the contact matrix,
demography data, and susceptibility matrix for solvers.
}
