% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_spread.R
\name{epi_spread}
\alias{epi_spread}
\title{Prepare contact and demography data for multiple risk groups}
\usage{
epi_spread(contact_matrix, demography_vector, p_susceptibility, susceptibility)
}
\arguments{
\item{contact_matrix}{Social contact matrix. Entry $mm_{ij}$ gives average
number of contacts in group $i$ reported by participants in group j}

\item{demography_vector}{Demography vector. Entry $pp_{i}$ gives proportion
of total population in group $i$ (model will normalise if needed)}

\item{p_susceptibility}{A matrix giving the probability that an individual
in demography group $i$ is in risk (or susceptibility) group $j$.
Each row represents the overall distribution of individuals in demographic
group $i$ across risk groups, and each row *must sum to 1.0*.}

\item{susceptibility}{A matrix giving the susceptibility of individuals in
demographic group $i$ and risk group $j$.}
}
\value{
A list object with named elements: "contact_matrix",
"demography_vector", "p_susceptibility_", and "susceptibility".
The contact matrix is replicated row and column wise for each risk group
and the demography vector is replicated for each risk group.
}
\description{
Prepare contact and demography data for multiple risk groups
}
