% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{final_size_}
\alias{final_size_}
\title{Final size of an epidemic}
\usage{
final_size_(
  contact_matrix,
  demography_vector,
  p_susceptibility,
  susceptibility,
  solver = "iterative",
  iterations = 10000L,
  tolerance = 1e-06,
  step_rate = 1.9,
  adapt_step = TRUE
)
}
\arguments{
\item{contact_matrix}{Social contact matrix. Entry \eqn{m_{ij}} gives the
average number of contacts in group \eqn{i} reported by participants in
group \eqn{j}.}

\item{demography_vector}{Demography vector. Entry \eqn{v_{i}} gives proportion
of total population in group \eqn{i} (model will normalise if needed)}

\item{p_susceptibility}{A matrix giving the probability that an individual
in demography group \eqn{i} is in risk (or susceptibility) group \eqn{j}.
Each row represents the overall distribution of individuals in demographic
group \eqn{i} across risk groups, and each row \emph{must sum to 1.0}.}

\item{susceptibility}{A matrix giving the susceptibility of individuals in
demographic group \eqn{i} and risk group \eqn{j}.}

\item{solver}{Which solver to use. Options are "iterative" or "newton", for
the iterative solver, or the Newton solver, respectively. The Newton solver
only uses the \code{iterations} and \code{tolerance} options.}

\item{iterations}{The number of iterations over which to solve for the final
size, unless the error is below the solver tolerance.}

\item{tolerance}{The solver tolerance, set to \code{1e-6} by default; solving for
final size ends when the error drops below this tolerance.}

\item{step_rate}{The solver step rate. Defaults to 1.9 as a value found to
work well.}

\item{adapt_step}{Boolean, whether the solver step rate should be changed
based on the solver error. Defaults to TRUE.}
}
\description{
Calculates the final size of an epidemic outbreak
in a population with heterogeneous mixing, and with heterogeneous
susceptibility to infection such as that conferred by an immunisation
programme.

\strong{Note}: This is a semi-internal function that is called by
\link[finalsize]{final_size}. This function is somewhat faster, but lacks
argument checks found in \code{final_size}. Use with caution.
}
\keyword{epidemic}
\keyword{model}
