% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{final_size_grps_cpp}
\alias{final_size_grps_cpp}
\title{Calculate final epidemic size over risk groups with RcppEigen backend}
\usage{
final_size_grps_cpp(
  contact_matrix,
  demography_vector,
  p_susceptibility,
  susceptibility,
  iterations = 1000L,
  adapt_step = TRUE,
  tolerance = 1e-06
)
}
\arguments{
\item{contact_matrix}{Social contact matrix. Entry $mm_{ij}$ gives average
number of contacts in group $i$ reported by participants in group j}

\item{demography_vector}{Demography vector. Entry $pp_{i}$ gives proportion
of total population in group $i$ (model will normalise if needed)}

\item{p_susceptibility}{WIP.}

\item{susceptibility}{WIP.}

\item{iterations}{WIP.}

\item{adapt_step}{WIP}

\item{tolerance}{WIP.}
}
\description{
This function calculates final epidemic size using SIR model
for a heterogeneously mixing population, with risk groups
}
\keyword{epidemic}
\keyword{model}
