% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{final_size_newton_cpp}
\alias{final_size_newton_cpp}
\title{Cpp calculation of final outbreak size.}
\usage{
final_size_newton_cpp(
  contact_matrix,
  demography,
  susceptibility,
  adapt_step = TRUE,
  step_rate = 1.9,
  cache_pi = TRUE,
  tolerance = 1e-06
)
}
\arguments{
\item{contact_matrix}{Social contact matrix. Entry mm_ij gives average
number of contacts in group i reported by participants in group j.}

\item{demography}{Demography vector. Entry pp_i gives proportion of
total population in group i (model will normalise if needed)}

\item{susceptibility}{Proportion of each group susceptible. Null assumption is
fully susceptible.}

\item{adapt_step}{Solver step size is adaptive? WIP.}

\item{step_rate}{Solver step rate. WIP}

\item{cache_pi}{Whether to cache matrix? WIP.}

\item{tolerance}{WIP.}
}
\value{
Final size of the epidemic as a vector, with one element per age and susceptibility group.
}
\description{
Test function to calculate squares.
}
\keyword{epidemic}
\keyword{model}
