% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalsize-package.R
\docType{package}
\name{finalsize-package}
\alias{finalsize}
\alias{finalsize-package}
\title{finalsize: Calculate the Final Size of an Epidemic}
\description{
Calculate the final size of a susceptible-infectious-recovered epidemic in a population with demographic variation in contact patterns and susceptibility to disease, as discussed in Miller (2012) \doi{10.1007/s11538-012-9749-6}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/finalsize}
  \item \url{https://epiverse-trace.github.io/finalsize/}
  \item Report bugs at \url{https://github.com/epiverse-trace/finalsize/issues}
}

}
\author{
\strong{Maintainer}: Rosalind Eggo \email{r.eggo@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-0362-6717}{ORCID}) [contributor]

Authors:
\itemize{
  \item Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [copyright holder]
  \item Edwin Van Leeuwen \email{edwin.vanleeuwen@ukhsa.gov.uk} (\href{https://orcid.org/0000-0002-2383-5305}{ORCID}) [copyright holder]
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [contributor]
  \item Thibaut Jombart \email{thibaut@data.org} (\href{https://orcid.org/0000-0003-3796-2097}{ORCID}) [contributor]
  \item Andree Valle-Campos \email{andree.valle-campos@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-7779-481X}{ORCID}) [contributor]
  \item Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [reviewer]
}

}
\keyword{internal}
