% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalSize.R
\name{FinalSize}
\alias{FinalSize}
\title{Final size analysis}
\usage{
FinalSize(x, s0, i0, alpha = 0.05, onesided = FALSE, max.val = 250)
}
\arguments{
\item{x}{Number of cases at the end of the outbreak (single value or vector)}

\item{s0}{Number of initial susceptibles (single value or vector)}

\item{i0}{Number of initial infectious (single value or vector)}

\item{alpha}{Significance level default = 0.05}

\item{onesided}{Test onesided default = FALSE}

\item{max.val}{Maximum value of R for numeric optimization}

\item{comp}{Direction of comparison i.e.(`>`or`<`)#' @param i0}
}
\value{
Table with estimator, confidence interval and test R<1, R>1 and R = 1
 point.est     ci.ll     ci.ul pval.above1 pval.below1 pval.equal1
}
\description{
Final size analysis
}
\examples{
#tests for same code as Mathematica book of Mart de Jong
FinalSize(c(0,0,0,0),c(2,2,2,2),c(2,2,2,2),onesided = T)
FinalSize(c(2,2,2,2),c(2,2,2,2),c(2,2,2,2),onesided = T)
FinalSize(c(2,1,1,0),c(2,2,2,2),c(2,2,2,2))
FinalSize(c(3),c(20),c(20),max.val = 50)

FinalSize(c(7,4),c(40,19)-1, c(1,1))
FinalSize(c(5,5),c(40,40)-1, c(1,1))
FinalSize(c(7),c(40)-1, c(1))
FinalSize(c(6,2),c(24,12)-1, c(1,1))

}
