% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalSize.R
\name{pFS}
\alias{pFS}
\title{probability of final size given R}
\usage{
pFS(R, x, s0, i0)
}
\arguments{
\item{R}{Value of R0}

\item{x}{Number of cases at the end of the outbreak}

\item{s0in}{Number of initial susceptibles}

\item{i0in}{Number of initial infectious}

\item{r0in}{Number of initial recovered}
}
\value{
probability to observe x given other values
}
\description{
probability of final size given R
}
\examples{
pFS(2,0,1,1)
pFS(2,1,1,1)

}
