% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.solve_newton}
\alias{.solve_newton}
\title{Newton solver for final size.}
\usage{
.solve_newton(
  contact_matrix,
  demography_vector,
  susceptibility,
  iterations = 10000L,
  tolerance = 1e-06
)
}
\arguments{
\item{contact_matrix}{Social contact matrix. Entry $mm_{ij}$ gives average
number of contacts in group $i$ reported by participants in group j}

\item{demography_vector}{Demography vector. Entry $pp_{i}$ gives proportion
of total population in group $i$ (model will normalise if needed)}

\item{susceptibility}{A matrix giving the susceptibility of individuals in
demographic group $i$ and risk group $j$.}

\item{iterations}{Number of solver iterations. Defaults to 1,000.}

\item{tolerance}{Solver error tolerance.}
}
\value{
A vector final sizes, of the size (N demography groups *
N risk groups).

A two dimensional array of final sizes per age-risk group.
}
\description{
Newton solver for final size.
}
