% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{final_size_cpp}
\alias{final_size_cpp}
\title{Calculate final epidemic size with RcppEigen backend}
\usage{
final_size_cpp(
  r0,
  contact_matrix,
  demography_vector,
  prop_initial_infected,
  prop_suscep,
  iterations = 30L
)
}
\arguments{
\item{r0}{Basic reproduction number.}

\item{contact_matrix}{Social contact matrix. Entry $mm_{ij}$ gives average
number of contacts in group $i$ reported by participants in group j}

\item{demography_vector}{Demography vector. Entry $pp_{i}$ gives proportion
of total population in group $i$ (model will normalise if needed)}

\item{prop_initial_infected}{Proportion of all age groups that is initially
infected. May be a single number, or a vector of proportions infected.
If a vector, must be the same length as the demography vector, otherwise the
vector will be recycled. Default value is 0.001 for all groups.}

\item{prop_suscep}{Proportion of each group susceptible. May be a single
numeric value or a numeric vector of the same length as the demography
vector.}

\item{iterations}{Number of solver iterations}
}
\description{
This function calculates final epidemic size using SIR model 
for a heterogeneously mixing population.
}
\keyword{epidemic}
\keyword{model}
