% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{final_size_grps_cpp}
\alias{final_size_grps_cpp}
\title{Calculate final epidemic size with risk groups using Eigen backend}
\usage{
final_size_grps_cpp(
  contact_matrix,
  demography_vector,
  p_susceptibility,
  susceptibility,
  solver,
  iterations = 1000L,
  tolerance = 1e-06,
  step_rate = 1.9,
  adapt_step = TRUE
)
}
\arguments{
\item{contact_matrix}{Social contact matrix. Entry $mm_{ij}$ gives average
number of contacts in group $i$ reported by participants in group j}

\item{demography_vector}{Demography vector. Entry $pp_{i}$ gives proportion
of total population in group $i$ (model will normalise if needed)}

\item{p_susceptibility}{A matrix giving the probability that an individual
in demography group $i$ is in risk (or susceptibility) group $j$.
Each row represents the overall distribution of individuals in demographic
group $i$ across risk groups, and each row *must sum to 1.0*.}

\item{susceptibility}{A matrix giving the susceptibility of individuals in
demographic group $i$ and risk group $j$.}

\item{solver}{Which solver to use. Options are "iterative" or "newton", for
the iterative solver, or the Newton solver, respectively. Special conditions
apply when using the Newton solver.}

\item{iterations}{Number of solver iterations. Defaults to 1,000.}

\item{tolerance}{Solver error tolerance.}

\item{step_rate}{The solver step rate for the iterative solver. Defaults to
1.9 as a value found to work well.}

\item{adapt_step}{Boolean, whether the iterative solver step rate should be
changed based on the solver error. Defaults to TRUE.}
}
\description{
This function calculates final epidemic size using SIR model
for a heterogeneously mixing population, using an Eigen backend
}
\keyword{epidemic}
\keyword{model}
