% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r0_conversions.R
\name{lambda_to_r0}
\alias{lambda_to_r0}
\title{Calculate \eqn{R_0} from transmission rate (\eqn{\lambda})}
\usage{
lambda_to_r0(
  lambda,
  contact_matrix,
  demography_vector,
  infectious_period = 1.8
)
}
\arguments{
\item{lambda}{The transmission rate of the disease, also called the 'force of
infection' (\eqn{\lambda}). This is different from the effective
transmission rate (\eqn{\beta}).}

\item{contact_matrix}{Social contact matrix. Entry \eqn{m_{ij}} gives average
number of contacts in group \eqn{i} reported by participants in group \eqn{j}}

\item{demography_vector}{Demography vector. Entry \eqn{v_{i}} gives
proportion of total population in group \eqn{i}.}

\item{infectious_period}{Duration of the infectious period in days.
Default value is 1.8 days.}
}
\value{
Returns the \eqn{R_0} for the infection in the population.
}
\description{
Uses the transmission rate (\eqn{\lambda}), contact matrix
(\eqn{c}), population (\eqn{N}), and infectious period (\eqn{\gamma}) to
calculate the R0 using the following equation.
\deqn{\lambda max(EV(C)) \gamma}
where \eqn{EV(C)} denotes the eigenvalues of the matrix \eqn{C} which is
calculated from the contact matrix and the population
(\eqn{C[i,j] = c[i,j] N[j]}).
}
\examples{
# Get example dataset and prepare contact matrix and demography
data(polymod_uk)
contact_matrix <- polymod_uk$contact_matrix
demography_vector <- polymod_uk$demography_vector

# define lambda
lambda <- 0.3

# define infectious period of 5 days
infectious_period <- 5

lambda_to_r0(
  lambda, contact_matrix, demography_vector, infectious_period
)

}
